update(add): added emtpy views if no items found

+++ todoList/Views/ListView.swift
@@ -11,19 +11,25 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) {item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateNoteCompletion(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                EmptyItemsView()
+            } else {
+                List {
+                    ForEach(listViewModel.items) {item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateNoteCompletion(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(.plain)
         .navigationTitle("Todo List 📝")
         .toolbar {
             ToolbarItem(placement: .topBarLeading, content: {
@@ -48,3 +54,4 @@ struct ListView: View {
     }
     .environmentObject(ListViewModel())
 }
+
